breaking out the 'waitUntilNoRequests'
