+++ test/e2e/doc-list.spec.js
@@ -1,3 +1,5 @@
+const wsHelper = require('./test-helper');
+
 const host = process.env.CI === 'true' ? 'localhost' : 'host.docker.internal';
 const engine = process.env.CI === 'true' ? 'localhost' : 'qix-engine';
 
@@ -8,54 +10,13 @@ const OPTS = {
 };
 let page;
 
-const idArray = [];
-let lastRequest;
-
-const setIntervalWS = (callback, delay) => {
-  const intervalID = setInterval(() => {
-    callback();
-  }, delay);
-  return intervalID;
-};
-
-const waitUntilNoRequests = idleTime => new Promise((resolve) => {
-  const intervalID = setIntervalWS(() => {
-    const silence = new Date().getTime() - lastRequest;
-    // console.log('Silence', silence, idArray.length === 0, silence > idleTime);
-    if (idArray.length === 0 && silence > idleTime) {
-      clearInterval(intervalID);
-      resolve();
-    }
-  }, 100);
-});
-
 describe('doc-list', () => {
   beforeEach(async () => {
     page = await browser.newPage();
     const client = page._client;
     await client.send('Animation.setPlaybackRate', { playbackRate: 12 });
 
-    client.on('Network.webSocketFrameSent', ({ /* requestId, timestamp, */ response }) => {
-      // console.log('Network.webSocketFrameSent', requestId, timestamp, response.payloadData);
-      const sentJSON = JSON.parse(response.payloadData);
-      // console.log('Network.webSocketFrameSent', sentJSON);
-      idArray.push(sentJSON.id);
-      lastRequest = new Date().getTime();
-      // console.log('Network.webSocketFrameSent', idArray, lastRequest);
-    });
-
-    client.on('Network.webSocketFrameReceived', ({ /* requestId, timestamp, */ response }) => {
-      // console.log('Network.webSocketFrameReceived', requestId, timestamp, response.payloadData);
-      const receivedJSON = JSON.parse(response.payloadData);
-      // console.log('Network.webSocketFrameReceived', receivedJSON.id);
-
-      const index = idArray.indexOf(receivedJSON.id);
-      if (index > -1) {
-        idArray.splice(index, 1);
-      }
-
-      // console.log('Network.webSocketFrameReceived', idArray);
-    });
+    wsHelper.init(client);
   });
 
   it('should show the no engine found with invalid engine url', async () => {
@@ -79,7 +40,7 @@ describe('doc-list', () => {
       page.waitForNavigation({ waitUntil: 'domcontentloaded' }),
     ]);
 
-    await waitUntilNoRequests(500);
+    await wsHelper.waitUntilNoRequests(500);
 
     // This seems to be the last elements to render, the text inside the columns.
     await page.waitForSelector('.name-and-text > .bartext', { visible: true });
